# docs/swagger.yaml
basePath: /
definitions:
...
  request.RondaAttendanceCreateRequest:
    properties:
      note:
        type: string
      resident_id:
        type: string
      ronda_activity_id:
        type: string
      status:
        description: hadir / tidak_hadir
        type: string
    type: object
  request.RondaAttendanceUpdateRequset:
    properties:
      id:
        type: string
      note:
        type: string
      resident_id:
        type: string
      ronda_activity_id:
        type: string
      status:
        description: hadir / tidak_hadir
        type: string
    type: object
  request.RondaConstributionCreateRequest:
    properties:
      date:
        type: string
      name:
        type: string
      ronda_group_id:
        type: string
      total:
        type: number
      total_penalty:
        type: number
    type: object
  request.RondaConstributionUpdateRequset:
    properties:
      date:
        type: string
      id:
        type: string
      name:
        type: string
      ronda_group_id:
        type: string
      total:
        type: number
      total_penalty:
        type: number
    type: object
...
  utils.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      sort:
        type: string
      sort_by:
        type: string
      total_pages:
        type: integer
      total_rows:
        type: integer
    type: object
  utils.ResponseData:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Pagination'
      status:
        type: string
    type: object
host: localhost:5001
info:
  contact:
    email: sarikhin@yahoo.co.id
    name: API Support
  description: This is a API Server for SIRNAWA
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SIRNAWA API
  version: "1.0"
paths:
...
  /ronda-attendance:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Get List RondaAttendance
      tags:
      - RondaAttendance
    post:
      consumes:
      - application/json
      parameters:
      - description: Create RondaAttendance
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RondaAttendanceCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Create RondaAttendance
      tags:
      - RondaAttendance
  /ronda-attendance/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: RondaAttendance id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Delete RondaAttendance By ID
      tags:
      - RondaAttendance
    get:
      consumes:
      - application/json
      parameters:
      - description: RondaAttendance id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Find RondaAttendance By ID
      tags:
      - RondaAttendance
    put:
      consumes:
      - application/json
      parameters:
      - description: Update RondaAttendance
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RondaAttendanceUpdateRequset'
      - description: RondaAttendance id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Update RondaAttendance
      tags:
      - RondaAttendance
  /ronda-attendance/paginated:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Order by
        in: query
        name: order_by
        type: string
      - description: Order
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Get Paginated RondaAttendance
      tags:
      - RondaAttendance
  /ronda-constribution:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Get List RondaConstribution
      tags:
      - RondaConstribution
    post:
      consumes:
      - application/json
      parameters:
      - description: Create RondaConstribution
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RondaConstributionCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Create RondaConstribution
      tags:
      - RondaConstribution
  /ronda-constribution/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: RondaConstribution id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Delete RondaConstribution By ID
      tags:
      - RondaConstribution
    get:
      consumes:
      - application/json
      parameters:
      - description: RondaConstribution id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Find RondaConstribution By ID
      tags:
      - RondaConstribution
    put:
      consumes:
      - application/json
      parameters:
      - description: Update RondaConstribution
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RondaConstributionUpdateRequset'
      - description: RondaConstribution id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Update RondaConstribution
      tags:
      - RondaConstribution
  /ronda-constribution/paginated:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Order by
        in: query
        name: order_by
        type: string
      - description: Order
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Get Paginated RondaConstribution
      tags:
      - RondaConstribution
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"